Moved Python bindings to dedicated repository: https://github.com/spaceinventor/PyCSH
The Python bindings started out with the sole purpose of allowing access to the parameter system from Python,
but as their interface and functionality eventually grew to mimic that of CSH,
calling them libparam bindings started to become misleading.

This, combined with the fact that they could be quite tricky to configure, compile and import for dedicated project,
has provided ample reasoning to move them to a dedicated repository.

They are now quite a bit easier to embed into projects.
It's also possible that they may eventually be uploaded to PyPi,
now that they are structured in a way which is far more friendly to the idea.