Accept other nodeid's on push
We accept the fact that a packet can contain a parameter with the nodeid = 0, for push meaning it should be saved to local. And for pull meaning it should be gotten from local. The client will rewrite the nodeid to the source address, and the server will rewrite the nodeid to = 0.

This enables the wide reuse of parameters across multiple nodes, and even brodcast.

it's a dangerous feature and it needs more testing. But it's necessary for CSP recovery.