vmem client abort
Abort any transfer that is ongoing
This function can be called from outside the task that is currently running a transfer
The transferring task will monitor the abort flag and then stop on the next coming
opportunity.

The flag is not thread safe and will thus abort all transfers.
The flag is reset at the next call to up or download