Introduced a VMEM request type plug-in mechanism to be used to extend the VMEM server with new features which not necessarily belongs to VMEM in public domain. (#45)
Introduced a VMEM request type plug-in mechanism to be used to extend the VMEM server with new features which not necessarily belongs to VMEM in public domain.

Now one can use the vmem_server_bind_type() method for registering a handler function which will be called in the VMEM server context for a particular VMEM request. These VMEM requests Are all request ID's with bit-7 set to '1' to indicate USER request.

The handler will be registered in a linked list which will be maintained by the VMEM server it self. The memory used for each handler must be handled by the calling module (user), so that the VMEM server does not require configuration prior to compilation.